chore: Move uuid function to test/utils
