+++ test/utils/subtaskGenerator.ts
@@ -1,4 +1,5 @@
-import { SubtaskSchema, uuid, TeambitionTypes } from '../index'
+import { SubtaskSchema, TeambitionTypes } from '../index'
+import { uuid } from './uuid'
 import { random, randomNumber } from './random'
 export default function (limit: number, taskId: TeambitionTypes.TaskId) {
   const result: SubtaskSchema[] = []
